+++ services/devicemanagerservice/src/authentication/auth_message_processor.cpp
@@ -35,7 +35,7 @@ std::vector<std::string> AuthMessageProcessor::CreateAuthRequestMessage()
 {
     LOGI("AuthMessageProcessor::CreateAuthRequestMessage start.");
     std::vector<std::string> jsonStrVec;
-    int32_t thumbnailSize = authRequestContext_->appThumbnail.size();
+    int32_t thumbnailSize =(int32_t)(authRequestContext_->appThumbnail.size());
     int32_t thumbnailSlice = ((thumbnailSize / MSG_MAX_SIZE) + (thumbnailSize % MSG_MAX_SIZE) == 0 ? 0 : 1);
     nlohmann::json jsonObj;
     jsonObj[TAG_VER] = DM_ITF_VER;
@@ -69,7 +69,7 @@ std::vector<std::string> AuthMessageProcessor::CreateAuthRequestMessage()
 
         int32_t leftLen = thumbnailSize - idx * MSG_MAX_SIZE;
         int32_t sliceLen = (leftLen > MSG_MAX_SIZE) ? MSG_MAX_SIZE : leftLen;
-        LOGI("TAG_APP_THUMBNAIL encode, idx %d, sliceLen %d, thumbnailSize %d", idx, sliceLen, thumbnailSize);
+        LOGI("TAG_APP_THUMBNAIL encode, idx %d, sliceLen %d, thumbnailSize %d", idx, (uint32_t)sliceLen, thumbnailSize);
         jsonObj[TAG_APP_THUMBNAIL] = authRequestContext_->appThumbnail.substr(idx * MSG_MAX_SIZE, sliceLen);
         jsonStrVec.push_back(jsonThumbnailObj.dump());
     }
